fix: wrap Timeline legend to multiple lines instead of horizontal scroll
Extract FlowLayout to shared file and replace horizontal ScrollView
with wrapping flow layout so all legend entries are visible.

Closes #16
